some cleanups; logging refactored into helpers
